# [corefoundation] Optimize `CFArray`