--- a/tests/perftest/legacy/perftest-legacy.csproj
+++ b/tests/perftest/legacy/perftest-legacy.csproj
@@ -15,6 +15,7 @@
     <MonoBundlingExtraArgs>--registrar:static</MonoBundlingExtraArgs>
     <Optimize>true</Optimize> <!-- always optimize -->
     <LinkMode>None</LinkMode> <!-- never link -->
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>