@vs-mobiletools-engineering-service2 2021-09-15
### :x: [PR Build] Tests failed on Build :x:

Tests failed on Build. 
* [Azure DevOps](https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=5208833&view=ms.vss-test-web.test-result-details)
* [Html Report (VSDrops)](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/sim/;/tests/vsdrops_index.html)
# API diff
:white_check_mark: API Diff from stable
<details><summary>View API diff</summary>

* iOS [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/apidiff/;/ios-api-diff.html) [gist](https://gist.github.com/132dbad13e3ddb385ec0143eb4ac72d8)
* macOS [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/apidiff/;/mac-api-diff.html) [gist](https://gist.github.com/b8ab964cd46099503718dd305996126b)
* macCat [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/apidiff/;/maccat-api-diff.html) [gist](https://gist.github.com/5be55745efcbcd2b217d6a53c85641d1)
* macCatiOS  [gist](https://gist.github.com/6746890f88c7bf6e53dd0bbfc78d1a5f)
* tvOS [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/apidiff/;/tvos-api-diff.html) [gist](https://gist.github.com/36d93a6ec5b6798111cbba28cf0db7e6)
* watchOS [vsdrops](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/apidiff/;/watchos-api-diff.html) [gist](https://gist.github.com/2416b12df0ddeb33a030d1ceec79807f)
</details>

# API & Generator diff

:information_source: [API Diff (from PR only)](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/apigeneratordiff/;//api-diff.html) (please review changes)
:white_check_mark: [Generator Diff](https://vsdrop.corp.microsoft.com/file/v1/xamarin-macios/device-tests/20210915.26/5208833/apigeneratordiff/;//generator-diff/index.html) (only version changes)

# GitHub pages

Results can be found in the following github pages (it might take some time to publish):

* [Test results](https://xamarin.github.io/macios.ci/pr/PR12740/5208833/HtmlReport-sim/tests/vsdrops_index.html)
* [API diff ](https://xamarin.github.io/macios.ci/pr/PR12740/5208833/HtmlReport-sim/api-diff/api-diff.html)
* [API & Generator diff](https://xamarin.github.io/macios.ci/pr/PR12740/5208833/apicomparison/api-diff.html)


# Test results

<details>
<summary>4 tests failed, 98 tests passed.</summary>

## Failed tests

 * introspection/Mac Catalyst [dotnet]/Debug [dotnet]: Failed (Tests run: 44 Passed: 40 Inconclusive: 0 Failed: 1 Ignored: 3)
 * fsharp/Mac [dotnet]/Debug [dotnet]: TimedOut (Execution timed out after 1200 seconds.
No test log file was produced)
 * link sdk/tvOS - simulator/Debug: Failed
 * link sdk/tvOS - simulator/Release: Failed
</details>


[Pipeline](https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=5208833&view=ms.vss-test-web.test-result-details) on Agent XAMBOT-1096.BigSur'
Merge 143ce45a5596019c5777423f1ac69213a2bde8fb into b9cde5eeea898ef28663ea939b98fb85c9cf41ca