@rolfbjarne 2021-09-16
Test failures are unrelated

* introspection/Mac Catalyst [dotnet]/Debug [dotnet]: https://github.com/xamarin/xamarin-macios/issues/12734
* fsharp/Mac [dotnet]/Debug [dotnet]: #12738
* link sdk/tvOS - simulator/Debug: https://github.com/xamarin/xamarin-macios/issues/12640
* link sdk/tvOS - simulator/Release: https://github.com/xamarin/xamarin-macios/issues/12640